Fixed the decode()'s long/int bug

+++ src/com/kennesaw/cpumodule/Instruction.java
@@ -13,8 +13,8 @@ public class Instruction {
     public Instruction() {
     }
 
-    public void decodeInstruction(long instructionBin) {
-        String binStr = Long.toBinaryString(instructionBin);
+    public void decodeInstruction(int instructionBin) {
+        String binStr = Integer.toBinaryString(instructionBin);
         while (binStr.length() < 32) binStr = "0" + binStr;
         format = Byte.parseByte(binStr.substring(0,2), 2);
         opcode = Byte.parseByte(binStr.substring(2,8), 2);
